Added compression script + asm files for dialogues.
The ASM files don't reflect the bugfixes lory90 did on the
-uncomp project. The reason for this is that I wanted something
that did a clean rebuild of the dialogue.

Unfortunately that's not 100% the case. I can confirm that
once uncompressed, the bin files end up creating the same
data as the data created by the compress_script.py before
compression. It's byte-by-byte identical. However, my koscmp
program creates smaller .bin files from that uncompressed
source. The files do appear to work properly in the ROM
so I don't think that breaks anything.

However, for that reason, the .bin files remain the original
ones I extracted.

See documentation in script/documentation.txt on usage of
the script compression utility.