Fix dialogue saving.
Basically when dialogs are uncompressed, we need to ensure the
dialogue pointer that is saved is not a mere word, because it's
not in RAM anymore. Unfortunately we can't use the original
location. Fortunately, lory1990 had fixed this in the other
branch, so we reset the Saved_Dialogue_Addr to a different RAM
address so we can store a double word.

Also provide a couple of macros so we don't end up with a
zillion if/else/endif in the code. pushdlg means save the dialogue
pointer from a0 to the RAM location, and popdlg means restore
from the RAM to a0 (trashing d0). We could probably afford not
to trash d0 in the uncompressed case but I'd rather do this
in case some code following the popdlg macro does expect the
value to be sitting in d0.